refactor: restructure test environment and remove deprecated files
