--- new_pytest_env/app.py
@@ -1,9 +1,0 @@
-def add(a, b):
-    """Returns the sum of two numbers."""
-    return a + b
-
-def subtract(a, b):
-    """Returns the difference between two numbers."""
-    print("ari")
-
-    return a - b
