+++ new_pytest_env/pytest.ini
@@ -1,7 +1,7 @@
-[pytest]
-asyncio_default_fixture_loop_scope = function
-log_cli = true
-log_level = INFO
-log_cli_level = INFO
-log_file = pytest_logs.log
-log_file_level = DEBUG
+# [pytest]
+# asyncio_default_fixture_loop_scope = function
+# log_cli = true
+# log_level = INFO
+# log_cli_level = INFO
+# log_file = pytest_logs.log
+# log_file_level = DEBUG
