+++ new_pytest_env/test_deploy.py
@@ -1,8 +1,8 @@
-import pytest
-import os
+# import pytest
+# import os
 
-def test_gitea_deployment(deploy):
-    #assert os.path.exists(deploy["clone_directory"])
-    #assert deploy["access_token"]
-    #print(f"Repository '{deploy['access_token']}' is ready for testing.")
-    pass
+# def test_gitea_deployment(deploy):
+#     #assert os.path.exists(deploy["clone_directory"])
+#     #assert deploy["access_token"]
+#     #print(f"Repository '{deploy['access_token']}' is ready for testing.")
+#     pass
