+++ tests/conftest.py
@@ -1,3 +1,6 @@
+import os
+import shutil
+import tempfile
 import threading
 import time
 import debugpy
@@ -5,21 +8,35 @@ import docker
 import pytest
 from testcontainers.core.waiting_utils import wait_for_logs
 from tests import utils
-from tests.broadcast_container import BroadcastContainer
-from tests.opal_client_container import OpalClientContainer
-from tests.opal_server_container import OpalServerContainer
+from tests.containers.broadcast_container import BroadcastContainer
+from tests.containers.gitea_container import GiteaContainer
+from tests.containers.opal_client_container import OpalClientContainer
+from tests.containers.opal_server_container import OpalServerContainer
 
 from . import settings as s
 
 # wait up to 30 seconds for the debugger to attach
 def cancel_wait_for_client_after_timeout():
-    time.sleep(30)
+    time.sleep(5)
     debugpy.wait_for_client.cancel()
 
 t = threading.Thread(target=cancel_wait_for_client_after_timeout)
 t.start()
 print("Waiting for debugger to attach... 30 seconds timeout")
 debugpy.wait_for_client()
+
+@pytest.fixture(scope="session")
+def temp_dir():
+    # Setup: Create a temporary directory
+    dir_path = tempfile.mkdtemp()
+    print(f"Temporary directory created: {dir_path}")
+    yield dir_path
+
+    # Teardown: Clean up the temporary directory
+    shutil.rmtree(dir_path)
+    print(f"Temporary directory removed: {dir_path}")
+
+repo_name = "opal-example-policy-repo"
 
 @pytest.fixture(scope="session")
 def opal_network():
@@ -34,10 +51,24 @@ def opal_network():
     network.remove()
     print("Network removed")    
 
+@pytest.fixture(scope="session")
+def gitea_server():
+    
+    with GiteaContainer(
+            GITEA_CONTAINER_NAME="test_container",
+            repo_name="test_repo",
+            temp_dir=os.path.join(os.path.dirname(__file__), "temp"),
+            data_dir=os.path.dirname(__file__),
+            gitea_base_url="http://localhost:3000"
+        ).with_network(net).with_network_aliases("gitea") as gitea_container: 
+        
+        gitea_container.deploy_gitea()
+        gitea_container.init_repo()
+        yield gitea_container
 
 @pytest.fixture(scope="session")
-def broadcast_channel(opal_network: str):
-    with BroadcastContainer(network=opal_network).with_network_aliases("broadcast_channel") as container:
+def broadcast_channel():
+    with BroadcastContainer() as container:
         yield container
 
 
@@ -57,8 +88,9 @@ def opal_server(opal_network: str, broadcast_channel: BroadcastContainer):
     ip_address = broadcast_channel.get_container_host_ip()
     exposed_port = broadcast_channel.get_exposed_port(5432)
 
-    opal_broadcast_uri = f"http://{ip_address}:{exposed_port}"
-
+    #opal_broadcast_uri = f"http://{ip_address}:{exposed_port}"
+    opal_broadcast_uri = f"postgres://test:test@broadcast_channel:5432"
+    
     with OpalServerContainer(network=opal_network, opal_broadcast_uri=opal_broadcast_uri).with_network_aliases("opal_server") as container:
 
         container.get_wrapped_container().reload()
